+++ pyas2lib/__init__.py
@@ -1,28 +1,24 @@
-from __future__ import absolute_import
-# from pyas2lib.as2 import DIGEST_ALGORITHMS, ENCRYPTION_ALGORITHMS,\
-#     MDN_CONFIRM_TEXT, MDN_FAILED_TEXT, Partner, Organization, Message, Mdn
-import sys
+from pyas2lib.as2 import DIGEST_ALGORITHMS
+from pyas2lib.as2 import ENCRYPTION_ALGORITHMS
+from pyas2lib.as2 import MDN_CONFIRM_TEXT
+from pyas2lib.as2 import MDN_FAILED_TEXT
+from pyas2lib.as2 import Mdn
+from pyas2lib.as2 import Message
+from pyas2lib.as2 import Organization
+from pyas2lib.as2 import Partner
 
-VERSION = (1, 0, 3)
+VERSION = (1, 1, 0)
 __version__ = '.'.join(map(str, VERSION))
 
 
 __all__ = [
     'VERSION',
-    # 'DIGEST_ALGORITHMS',
-    # 'ENCRYPTION_ALGORITHMS',
-    # 'MDN_CONFIRM_TEXT',
-    # 'MDN_FAILED_TEXT',
-    # 'Partner',
-    # 'Organization',
-    # 'Message',
-    # 'Mdn'
+    'DIGEST_ALGORITHMS',
+    'ENCRYPTION_ALGORITHMS',
+    'MDN_CONFIRM_TEXT',
+    'MDN_FAILED_TEXT',
+    'Partner',
+    'Organization',
+    'Message',
+    'Mdn'
 ]
-
-if (2, 7) <= sys.version_info < (3, 2):
-    # On Python 2.7 and Python3 < 3.2, install no-op handler to silence
-    # `No handlers could be found for logger "elasticsearch"` message per
-    # <https://docs.python.org/2/howto/logging.html#configuring-logging-for-a-library>
-    import logging
-    logger = logging.getLogger('pyas2lib')
-    logger.addHandler(logging.NullHandler())